requirements.txt
Flask==2.3.3
tldextract==5.1.2
scikit-learn==1.3.2
pandas==2.1.1
google-api-python-client==2.104.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.1.1
requests==2.31.0
joblib==1.3.2
Werkzeug==2.3.7
MarkupSafe==2.1.3